# services/api/docs/swagger.yaml
definitions:
...
  lib.Response:
    properties:
      message:
        description: response message
        type: string
      status:
        description: http status
        type: integer
    type: object
...
  model.CreateMuhasabahRequest:
    properties:
      content:
        type: string
      date:
        type: string
      is_private:
        type: boolean
      mood_score:
        maximum: 5
        minimum: 1
        type: integer
    required:
    - content
    - date
    type: object
  model.CreateNoteRequest:
    properties:
      content:
        maxLength: 5000
        type: string
      ref_id:
        type: integer
      ref_type:
        $ref: '#/definitions/model.NoteRefType'
    required:
    - content
    - ref_id
    - ref_type
    type: object
...
  model.NoteRefType:
    enum:
    - ayah
    - hadith
    type: string
    x-enum-varnames:
    - NoteRefTypeAyah
    - NoteRefTypeHadith
...
  model.RecordMurojaahRequest:
    properties:
      date:
        type: string
      duration_seconds:
        minimum: 0
        type: integer
      from_ayah:
        minimum: 1
        type: integer
      note:
        type: string
      score:
        maximum: 100
        minimum: 0
        type: integer
      surah_id:
        minimum: 1
        type: integer
      to_ayah:
        minimum: 1
        type: integer
    required:
    - date
    - surah_id
    type: object
...
  model.UpdateMuhasabahRequest:
    properties:
      content:
        type: string
      is_private:
        type: boolean
      mood_score:
        maximum: 5
        minimum: 0
        type: integer
    type: object
  model.UpdateNoteRequest:
    properties:
      content:
        type: string
    required:
    - content
    type: object
...
info:
  contact: {}
paths:
...
  /api/v1/muhasabah:
    get:
      parameters:
      - description: Limit (default 20)
        in: query
        name: limit
        type: integer
      - description: Offset (default 0)
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/lib.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/lib.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/lib.Response'
      summary: Get all muhasabah entries
      tags:
      - Personal
    post:
      consumes:
      - application/json
      parameters:
      - description: Muhasabah data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/model.CreateMuhasabahRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/lib.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/lib.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/lib.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/lib.Response'
      summary: Create muhasabah entry
      tags:
      - Personal
  /api/v1/muhasabah/{id}:
    delete:
      parameters:
      - description: Muhasabah ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/lib.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/lib.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/lib.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/lib.Response'
      summary: Delete muhasabah entry
      tags:
      - Personal
    get:
      parameters:
      - description: Muhasabah ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/lib.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/lib.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/lib.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/lib.Response'
      summary: Get muhasabah entry by ID
      tags:
      - Personal
    put:
      consumes:
      - application/json
      parameters:
      - description: Muhasabah ID
        in: path
        name: id
        required: true
        type: integer
      - description: Update data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/model.UpdateMuhasabahRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/lib.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/lib.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/lib.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/lib.Response'
      summary: Update muhasabah entry
      tags:
      - Personal
  /api/v1/murojaah/history:
    get:
      parameters:
      - description: Limit (default 20)
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/lib.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/lib.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/lib.Response'
      summary: Get murojaah history
      tags:
      - Belajar
  /api/v1/murojaah/result:
    post:
      consumes:
      - application/json
      parameters:
      - description: Murojaah result
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/model.RecordMurojaahRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/lib.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/lib.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/lib.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/lib.Response'
      summary: Submit murojaah result
      tags:
      - Belajar
  /api/v1/murojaah/session:
    get:
      parameters:
      - description: Number of ayahs (default 10)
        in: query
        name: count
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/lib.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/lib.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/lib.Response'
      summary: Get murojaah session (random ayahs)
      tags:
      - Belajar
  /api/v1/murojaah/stats:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/lib.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/lib.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/lib.Response'
      summary: Get murojaah stats
      tags:
      - Belajar
  /api/v1/notes:
    get:
      parameters:
      - description: Filter by reference type
        in: query
        name: ref_type
        type: string
      - description: Filter by reference ID
        in: query
        name: ref_id
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/lib.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/lib.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/lib.Response'
      summary: Get all notes
      tags:
      - Belajar
    post:
      consumes:
      - application/json
      parameters:
      - description: Note data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/model.CreateNoteRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/lib.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/lib.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/lib.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/lib.Response'
      summary: Create note
      tags:
      - Belajar
  /api/v1/notes/{id}:
    delete:
      parameters:
      - description: Note ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/lib.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/lib.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/lib.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/lib.Response'
      summary: Delete note
      tags:
      - Belajar
    put:
      consumes:
      - application/json
      parameters:
      - description: Note ID
        in: path
        name: id
        required: true
        type: integer
      - description: Update data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/model.UpdateNoteRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/lib.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/lib.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/lib.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/lib.Response'
      summary: Update note
      tags:
      - Belajar
...
swagger: "2.0"